LayoutTests/http/tests/misc/webtiming-ssl-expected.txt

This test checks that Web Timing reports secureConnectionStart correctly and has reasonable values for connectStart and connectEnd.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS timing.connectStart is >= timing.navigationStart
PASS timing.secureConnectionStart is >= timing.connectStart
PASS timing.connectEnd is >= timing.secureConnectionStart
PASS successfullyParsed is true

TEST COMPLETE
